+++ central/detection/lifecycle/manager.go
@@ -14,7 +14,6 @@ import (
 	baselineDataStore "github.com/stackrox/rox/central/processbaseline/datastore"
 	processDatastore "github.com/stackrox/rox/central/processindicator/datastore"
 	"github.com/stackrox/rox/central/reprocessor"
-	"github.com/stackrox/rox/central/sensor/service/connection"
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/logging"
 	"github.com/stackrox/rox/pkg/process/filter"
@@ -49,7 +48,7 @@ type Manager interface {
 func newManager(buildTimeDetector buildtime.Detector, deployTimeDetector deploytime.Detector, runtimeDetector runtime.Detector,
 	deploymentDatastore deploymentDatastore.DataStore, processesDataStore processDatastore.DataStore, baselines baselineDataStore.DataStore,
 	alertManager alertmanager.AlertManager, reprocessor reprocessor.Loop, deletedDeploymentsCache cache.DeletedDeployments, filter filter.Filter,
-	processAggregator aggregator.ProcessAggregator, connectionManager connection.Manager) *managerImpl {
+	processAggregator aggregator.ProcessAggregator) *managerImpl {
 	m := &managerImpl{
 		buildTimeDetector:       buildTimeDetector,
 		deployTimeDetector:      deployTimeDetector,
@@ -71,7 +70,6 @@ func newManager(buildTimeDetector buildtime.Detector, deployTimeDetector deployt
 
 		removedOrDisabledPolicies: set.NewStringSet(),
 		processAggregator:         processAggregator,
-		connectionManager:         connectionManager,
 	}
 
 	go m.flushQueuePeriodically()
