+++ central/detection/lifecycle/manager_impl_test.go
@@ -10,7 +10,6 @@ import (
 	alertManagerMocks "github.com/stackrox/rox/central/detection/alertmanager/mocks"
 	processBaselineDataStoreMocks "github.com/stackrox/rox/central/processbaseline/datastore/mocks"
 	reprocessorMocks "github.com/stackrox/rox/central/reprocessor/mocks"
-	connectionMocks "github.com/stackrox/rox/central/sensor/service/connection/mocks"
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/env"
 	"github.com/stackrox/rox/pkg/fixtures"
@@ -33,7 +32,6 @@ type ManagerTestSuite struct {
 	reprocessor                *reprocessorMocks.MockLoop
 	alertManager               *alertManagerMocks.MockAlertManager
 	deploymentObservationQueue *queueMocks.MockDeploymentObservationQueue
-	connectionManager          *connectionMocks.MockManager
 	manager                    *managerImpl
 	mockCtrl                   *gomock.Controller
 }
@@ -45,14 +43,12 @@ func (suite *ManagerTestSuite) SetupTest() {
 	suite.reprocessor = reprocessorMocks.NewMockLoop(suite.mockCtrl)
 	suite.alertManager = alertManagerMocks.NewMockAlertManager(suite.mockCtrl)
 	suite.deploymentObservationQueue = queueMocks.NewMockDeploymentObservationQueue(suite.mockCtrl)
-	suite.connectionManager = connectionMocks.NewMockManager(suite.mockCtrl)
 
 	suite.manager = &managerImpl{
 		baselines:                  suite.baselines,
 		reprocessor:                suite.reprocessor,
 		alertManager:               suite.alertManager,
 		deploymentObservationQueue: suite.deploymentObservationQueue,
-		connectionManager:          suite.connectionManager,
 	}
 }
 
@@ -83,8 +79,6 @@ func makeIndicator() (*storage.ProcessBaselineKey, *storage.ProcessIndicator) {
 		DeploymentId:  uuid.NewV4().String(),
 		ContainerName: uuid.NewV4().String(),
 		PodId:         uuid.NewV4().String(),
-		ClusterId:     uuid.NewV4().String(),
-		Namespace:     uuid.NewV4().String(),
 		Signal:        signal,
 	}
 	key := &storage.ProcessBaselineKey{
@@ -99,10 +93,10 @@ func makeIndicator() (*storage.ProcessBaselineKey, *storage.ProcessIndicator) {
 func (suite *ManagerTestSuite) TestBaselineNotFound() {
 	suite.T().Setenv(env.BaselineGenerationDuration.EnvVar(), time.Millisecond.String())
 	key, indicator := makeIndicator()
+	elements := fixtures.MakeBaselineItems(indicator.GetSignal().GetExecFilePath())
 	suite.baselines.EXPECT().GetProcessBaseline(gomock.Any(), key).Return(nil, false, nil)
-	suite.connectionManager.EXPECT().SendMessage(gomock.Any(), gomock.Any()).Return(nil).AnyTimes()
 	suite.deploymentObservationQueue.EXPECT().InObservation(key.GetDeploymentId()).Return(false).AnyTimes()
-	suite.baselines.EXPECT().UpdateProcessBaselineAndSetTimestamp(gomock.Any(), gomock.Any()).Return(nil)
+	suite.baselines.EXPECT().UpsertProcessBaseline(gomock.Any(), key, elements, true, true).Return(nil, nil)
 	_, err := suite.manager.checkAndUpdateBaseline(indicatorToBaselineKey(indicator), []*storage.ProcessIndicator{indicator})
 	suite.NoError(err)
 	suite.mockCtrl.Finish()
@@ -116,7 +110,7 @@ func (suite *ManagerTestSuite) TestBaselineNotFound() {
 
 	suite.mockCtrl = gomock.NewController(suite.T())
 	suite.baselines.EXPECT().GetProcessBaseline(gomock.Any(), key).Return(nil, false, nil)
-	suite.baselines.EXPECT().UpdateProcessBaselineAndSetTimestamp(gomock.Any(), gomock.Any()).Return(expectedError)
+	suite.baselines.EXPECT().UpsertProcessBaseline(gomock.Any(), key, elements, true, true).Return(nil, expectedError)
 	_, err = suite.manager.checkAndUpdateBaseline(indicatorToBaselineKey(indicator), []*storage.ProcessIndicator{indicator})
 	suite.Equal(expectedError, err)
 }
