+++ central/detection/lifecycle/singleton.go
@@ -14,7 +14,6 @@ import (
 	processDatastore "github.com/stackrox/rox/central/processindicator/datastore"
 	"github.com/stackrox/rox/central/processindicator/filter"
 	"github.com/stackrox/rox/central/reprocessor"
-	"github.com/stackrox/rox/central/sensor/service/connection"
 	"github.com/stackrox/rox/pkg/sync"
 	"github.com/stackrox/rox/pkg/utils"
 )
@@ -37,7 +36,6 @@ func initialize() {
 		cache.DeletedDeploymentsSingleton(),
 		filter.Singleton(),
 		aggregator.Singleton(),
-		connection.ManagerSingleton(),
 	)
 
 	policies, err := policyDataStore.Singleton().GetAllPolicies(lifecycleMgrCtx)
