+++ central/processbaseline/datastore/datastore_impl_test.go
@@ -16,7 +16,6 @@ import (
 	"github.com/stackrox/rox/pkg/fixtures/fixtureconsts"
 	"github.com/stackrox/rox/pkg/postgres"
 	"github.com/stackrox/rox/pkg/postgres/pgtest"
-	processBaselinePkg "github.com/stackrox/rox/pkg/processbaseline"
 	"github.com/stackrox/rox/pkg/protoassert"
 	"github.com/stackrox/rox/pkg/protocompat"
 	"github.com/stackrox/rox/pkg/sac"
@@ -72,7 +71,7 @@ func (suite *ProcessBaselineDataStoreTestSuite) TearDownTest() {
 }
 
 func (suite *ProcessBaselineDataStoreTestSuite) mustSerializeKey(key *storage.ProcessBaselineKey) string {
-	serialized, err := processBaselinePkg.KeyToID(key)
+	serialized, err := keyToID(key)
 	suite.Require().NoError(err)
 	return serialized
 }
@@ -308,9 +307,9 @@ func (suite *ProcessBaselineDataStoreTestSuite) TestIDToKeyConversion() {
 		Namespace:     "namespace",
 	}
 
-	id, err := processBaselinePkg.KeyToID(key)
+	id, err := keyToID(key)
 	suite.NoError(err)
-	resKey, err := processBaselinePkg.IDToKey(id)
+	resKey, err := IDToKey(id)
 	suite.NoError(err)
 	suite.NotNil(resKey)
 	protoassert.Equal(suite.T(), key, resKey)
