+++ pkg/features/list.go
@@ -134,9 +134,6 @@ var (
 
 	// Adds built-in policy to ensure that Red Hat images are signed by Red Hat Release Key
 	RedHatImagesSignedPolicy = registerFeature("Adds built-in policy to ensure that Red Hat images are signed by the Red Hat release key", "ROX_RED_HAT_IMAGES_SIGNED_POLICY", unchangeableInProd)
-
-	// AutolockAllProcessBaselines automatically locks all process baselines after the observation period
-	AutolockAllProcessBaselines = registerFeature("Automatically locks all process baselines after the observation period", "ROX_AUTOLOCK_ALL_PROCESS_BASELINES")
 )
 
 // The following feature flags are related to Scanner V4.
